refactor(player): migrate audio player to UiState and split composables
Refactor AudioPlayerViewModel to expose AudioPlayerUiState via StateFlow, replacing scattered composeState properties. Decompose AudioPlayerScreen into focused composables: AudioPlayerArtwork, AudioPlayerControls, AudioPlayerTopBar, AudioPlayerTrackInfo. Refactor MusicHomeViewModel to use MusicHomeUiState with type-safe section identifiers. Add sleep timer end-of-episode trigger in AudioPlaybackManager.
